Fix bug where for some percentile values was infinity cycling + Code maint.

+++ src/HistogramObject.cpp
@@ -33,7 +33,7 @@ double HistogramObject::getMax()
 	return this->max;
 }
 
-long HistogramObject::getFrequency()
+size_t HistogramObject::getFrequency()
 {
 	return this->frequency;
 }
